+++ src/core/pages/Snippet.ts
@@ -1,0 +1,18 @@
+import {Expr} from './Expr.js'
+import {ExprData} from './ExprData.js'
+
+export function snippet(
+  start = '<mark>',
+  end = '</mark>',
+  cutOff = '...',
+  limit = 64
+): Expr<string> {
+  return Expr(
+    ExprData.Call('snippet', [
+      ExprData.Value(start),
+      ExprData.Value(end),
+      ExprData.Value(cutOff),
+      ExprData.Value(limit)
+    ])
+  )
+}